pyproject.toml
[tool.poetry]
name = "hyperliquid-copy-trade-bot"
version = "0.1.0"
description = "Advanced CopyTrade bot for Hyperliquid with real-time leader tracking and risk management"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "copy_trade_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
hyperliquid-python-sdk = {path = "../hyperliquid-python-sdk"}
redis = "^5.0.0"
psycopg2-binary = "^2.9.0"
sqlalchemy = "^2.0.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
structlog = "^23.0.0"
aiohttp = "^3.8.0"
asyncio-mqtt = "^0.16.0"
websockets = "^11.0.0"
schedule = "^1.2.0"
python-dotenv = "^1.0.0"
click = "^8.1.0"
rich = "^13.0.0"
prometheus-client = "^0.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.0.0"
pre-commit = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
